+++ src/application/particle/EnemyBulletParticle.cpp
@@ -12,7 +12,7 @@
 #pragma comment(lib,"d3dcompiler.lib")
 
 ComPtr<ID3D12RootSignature>EnemyBulletParticle::rootsignature;
-ComPtr<ID3D12PipelineState>EnemyBulletParticle::pipelinestate;
+ComPtr<ID3D12PipelineState>EnemyBulletParticle::pipelinestate; 
 TextureManager* EnemyBulletParticle::spriteManager = nullptr;
 ID3D12Device* EnemyBulletParticle::device = nullptr;
 Camera* EnemyBulletParticle::camera = nullptr;
@@ -271,8 +271,8 @@ void EnemyBulletParticle::CreateBuffers()
 	D3D12_RESOURCE_DESC textureResourceDesc{};
 	textureResourceDesc.Dimension = D3D12_RESOURCE_DIMENSION_TEXTURE2D;
 	textureResourceDesc.Format = DXGI_FORMAT_R32G32B32A32_FLOAT;
-	textureResourceDesc.Width = textureWidth;	//幅
-	textureResourceDesc.Height = textureHeight;	//高さ
+	textureResourceDesc.Width = (UINT)textureWidth;	//幅
+	textureResourceDesc.Height = (UINT)textureHeight;	//高さ
 	textureResourceDesc.DepthOrArraySize = 1;
 	textureResourceDesc.MipLevels = 1;
 	textureResourceDesc.SampleDesc.Count = 1;
@@ -290,8 +290,8 @@ void EnemyBulletParticle::CreateBuffers()
 		0,
 		nullptr,
 		imageData,
-		sizeof(XMFLOAT4) * textureWidth,
-		sizeof(XMFLOAT4) * imageDataCount
+		sizeof(XMFLOAT4) * (UINT)textureWidth,
+		sizeof(XMFLOAT4) * (UINT)imageDataCount
 	);
 
 	//デスクリプタヒープ生成
