+++ src/application/particle/ExplosionParticle1.cpp
@@ -275,8 +275,8 @@ void ExplosionParticle1::CreateBuffers()
 	D3D12_RESOURCE_DESC textureResourceDesc{};
 	textureResourceDesc.Dimension = D3D12_RESOURCE_DIMENSION_TEXTURE2D;
 	textureResourceDesc.Format = DXGI_FORMAT_R32G32B32A32_FLOAT;
-	textureResourceDesc.Width = textureWidth;	//幅
-	textureResourceDesc.Height = textureHeight;	//高さ
+	textureResourceDesc.Width = (UINT)textureWidth;	//幅
+	textureResourceDesc.Height = (UINT)textureHeight;	//高さ
 	textureResourceDesc.DepthOrArraySize = 1;
 	textureResourceDesc.MipLevels = 1;
 	textureResourceDesc.SampleDesc.Count = 1;
@@ -294,8 +294,8 @@ void ExplosionParticle1::CreateBuffers()
 		0,
 		nullptr,
 		imageData,
-		sizeof(XMFLOAT4) * textureWidth,
-		sizeof(XMFLOAT4) * imageDataCount
+		sizeof(XMFLOAT4) * (UINT)textureWidth,
+		sizeof(XMFLOAT4) * (UINT)imageDataCount
 	);
 
 	//デスクリプタヒープ生成
