+++ src/engine/object/BillboardSpriteModel.cpp
@@ -23,7 +23,7 @@ void BillboardSpriteModel::CreateBuffers()
 	CreateVertex();
 
 	//頂点データ全体のサイズ
-	UINT sizeVB = static_cast<UINT>(sizeof(VertexPosNormalUv)) * vertices.size();
+	UINT sizeVB = static_cast<UINT>(sizeof(VertexPosNormalUv)) * (UINT)vertices.size();
 
 	//頂点バッファの設定
 	D3D12_HEAP_PROPERTIES heapProp{};	//ヒープ設定
@@ -130,8 +130,8 @@ void BillboardSpriteModel::CreateBuffers()
 	D3D12_RESOURCE_DESC textureResourceDesc{};
 	textureResourceDesc.Dimension = D3D12_RESOURCE_DIMENSION_TEXTURE2D;
 	textureResourceDesc.Format = DXGI_FORMAT_R32G32B32A32_FLOAT;
-	textureResourceDesc.Width = textureWidth;	//幅
-	textureResourceDesc.Height = textureHeight;	//高さ
+	textureResourceDesc.Width = (UINT)textureWidth;	//幅
+	textureResourceDesc.Height = (UINT)textureHeight;	//高さ
 	textureResourceDesc.DepthOrArraySize = 1;
 	textureResourceDesc.MipLevels = 1;
 	textureResourceDesc.SampleDesc.Count = 1;
@@ -149,8 +149,8 @@ void BillboardSpriteModel::CreateBuffers()
 		0,
 		nullptr,
 		imageData,
-		sizeof(XMFLOAT4) * textureWidth,
-		sizeof(XMFLOAT4) * imageDataCount
+		sizeof(XMFLOAT4) * (UINT)textureWidth,
+		sizeof(XMFLOAT4) * (UINT)imageDataCount
 	);
 
 	//デスクリプタヒープ生成
@@ -264,7 +264,7 @@ void BillboardSpriteModel::Update()
 
 	HRESULT result;
 	//頂点データ全体のサイズ
-	UINT sizeVB = static_cast<UINT>(sizeof(VertexPosNormalUv)) * vertices.size();
+	UINT sizeVB = static_cast<UINT>(sizeof(VertexPosNormalUv)) * (UINT)vertices.size();
 
 	//頂点バッファへのデータ転送
 	//GPU上のバッファに対応した仮想メモリ（メインメモリ上）を取得
